requirements.txt
psycopg2-binary==2.9.9
google-cloud-bigquery==3.23.0
google-cloud-storage>=2.10.0
pandas>=1.5.0
openpyxl>=3.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

